UMultiFab Interpolator defined
The interpolator is not optimized. And will go through several changes.
It is not yet clear whether the interpolator should be for class UMultiFab or for Array1D
Secondly current implementation is slow because each Real component of Array1D<Real,243> is being accessed individually, not as a whole.
This will be changed in the future.
Committing because the basic structure of custom interpolators has been defined.