Sum_moments_new introduced
Coarse particles close to the refinement level are interpolated to shared (coarse-fine) nodes using the fine-level interpolation function (shape function)
With equal-sized particles across levels, the error is totally eliminated
This is a local correction and does not need MPI , hence it is better than the ghost particle method.

The same methodology will be applied to jHat and NodeMM calculations next,

The fine level boundary is not marked on the coarse level NodeStatus, hence this method has a temporary function that marks the boundary and will be removed when NodeStatus is fixed

Methodology:
1. Let X be the Coarse Interpolation function in one direction, the fine interpolation function is (2X-1). But since we also divide by the volume of the cell, an additional multiplication
   by 2 is needed in all directions, So the final interpolation function is 2(2X-1)
2. This will yield negative values for particles outside the fine interpolation range, -ve values are set to zero.
3.  When we iterate over particles in a particular cell, we check each node. Shared nodes get the correction, while coarse-only nodes do not get the correction.

Notes:
1. If particles are not spaced equally, error is seen, but the magnitude is much smaller than the previous method
2. Using smaller particles in the fine region introduced error again, but it is not around the edge-fine interface but rather around the heavy-fine particle interface
and moves upwards as the particles move upwards with a constant velocity in our test.